Added react components
