+++ client/src/components/Facts.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Facts() {
+  return (
+    <div className="main-dash">Facts</div>
+  )
+}
+
+export default Facts
