+++ client/src/components/Quiz.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Quiz() {
+  return (
+    <div className="main-dash">Quiz</div>
+  )
+}
+
+export default Quiz
